ripenv/cli/pyproject.toml
[project]
name = "ripenv"
version = "0.1.0"
description = "Local-first CLI for encrypting environment secrets for ripenv projects"
readme = "README.md"
authors = [
    { name = "ripenv" }
]
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "pynacl>=1.5",
    "argon2-cffi>=23.1",
    "pydantic>=1.10",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "email-validator>=2.1"
]

[project.urls]
Homepage = "https://example.com/ripenv"

[project.scripts]
ripenv = "ripenv.main:app"

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = [
    "tests"
]
